+++ LobbyControl/src/Dependency/LobbyCompatibilityChecker.cs
@@ -1,0 +1,18 @@
+using System.Runtime.CompilerServices;
+using LobbyCompatibility.Enums;
+using LobbyCompatibility.Features;
+
+namespace LobbyControl.Dependency
+{
+    public static class LobbyCompatibilityChecker
+    {
+        public static bool Enabled { get { return BepInEx.Bootstrap.Chainloader.PluginInfos.ContainsKey("BMX.LobbyCompatibility"); } }
+        
+        [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
+        public static void Init()
+        {
+            PluginHelper.RegisterPlugin(LobbyControl.GUID, System.Version.Parse(LobbyControl.VERSION), CompatibilityLevel.ServerOnly, VersionStrictness.Minor);
+        }
+        
+    }
+}